+++ src/main/java/de/rabea/controller/MethodOptionsController.java
@@ -4,13 +4,13 @@ import de.rabea.Controller;
 import de.rabea.request.HttpRequest;
 import de.rabea.request.HttpVerb;
 import de.rabea.response.HttpResponse;
-import de.rabea.response.OptionsResponseHeader;
+import de.rabea.response.head.OptionsResponseHeader;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import static de.rabea.request.HttpVerb.*;
-import static de.rabea.response.StatusLine.OK;
+import static de.rabea.response.head.StatusLine.OK;
 
 public class MethodOptionsController extends Controller {
     private Map<HttpVerb, HttpResponse> responsesForMethods = new HashMap<>();
