+++ src/main/java/de/rabea/controller/ServerError.java
@@ -4,7 +4,7 @@ import de.rabea.Controller;
 import de.rabea.request.HttpRequest;
 import de.rabea.response.HttpResponse;
 
-import static de.rabea.response.StatusLine.SERVER_ERROR;
+import static de.rabea.response.head.StatusLine.SERVER_ERROR;
 
 public class ServerError extends Controller {
