+++ src/main/java/de/rabea/response/GetResponse.java
@@ -2,7 +2,7 @@ package de.rabea.response;
 
 import de.rabea.ContentStorage;
 import de.rabea.Controller;
-import de.rabea.controller.FormController;
+import de.rabea.response.head.StatusLine;
 
 public class GetResponse extends ResponseCreator {
