+++ src/main/java/de/rabea/response/PutResponse.java
@@ -1,5 +1,7 @@
 package de.rabea.response;
 
+import de.rabea.response.head.StatusLine;
+
 public class PutResponse extends ResponseCreator {
 
     private StatusLine statusLine;
