+++ src/main/java/de/rabea/response/ResponseBuilder.java
@@ -1,5 +1,7 @@
 package de.rabea.response;
 
+import de.rabea.response.head.StatusLine;
+
 public class ResponseBuilder {
 
     private final StatusLine statusLine;
